Separate withTheme from Button component
Separated withTheme from Button component.
Deleted ThemeProvider and unnecessary options from stories.
Updated test.
Added possibility to render into a.
Added external and href prop.
